test: test 대상 훅 변경
useCardBrandValidate -> useCardBrand
